backend/requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
dj-rest-auth==6.0.0
Django==5.0.4
django-allauth==0.61.1
django-environ==0.11.2
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-spectacular==0.27.2
environs==11.0.0
idna==3.7
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
marshmallow==3.21.1
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==24.0
pycparser==2.22
Pygments==2.17.2
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.0
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.4.4
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
